+++ CHANGELOG.md
@@ -18,9 +18,13 @@
 ### 💡 Enhancements 💡
 
 - Use OpenCensus `metric` package for process metrics instead of `stats` package (#5486)
+- Update OTLP to v0.18.0 (#5530)
 - Log histogram min/max fields with `logging` exporter (#5520)
 
 ### 🧰 Bug fixes 🧰
+
+- Update sum field of exponential histograms to make it optional (#5530)
+- Remove redundant extension shutdown call (#5532)
 
 ## v0.53.0 Beta